% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refflat.R
\name{determine_levels}
\alias{determine_levels}
\title{Determine levels}
\usage{
determine_levels(refflat, buffer = 0)
}
\arguments{
\item{refflat}{data frame. The input refflat data.}
}
\value{
data frame. refflat data with a level column added.
}
\description{
Determine y-axis levels for each gene, in case any overlap one another
}
