% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refflat.R
\name{draw_gene}
\alias{draw_gene}
\title{Draw a gene on a plot}
\usage{
draw_gene(gene_data, y, arrowhead_length = 0.125, angle = 15,
  lwd = 2)
}
\arguments{
\item{gene_data}{list. List representing the gene.}

\item{y}{numeric. The y coordinate at which to draw the gene.}

\item{arrowhead_length}{numeric. Length of the edges of the arrow head (in
inches).}

\item{angle}{numeric. Angle from the shaft of the arrow to the edge of the
arrow head.}

\item{lwd}{numeric. Weight of lines.}
}
\description{
Draw a gene on a plot
}
