% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refflat.R
\name{fetch_refflat_data}
\alias{fetch_refflat_data}
\title{Fetch refflat data}
\usage{
fetch_refflat_data(tmpdir = tempdir(), method = "auto")
}
\arguments{
\item{tmpdir}{character. A non-empty character vector giving the temporary
directory name (provided to tempfile()).}

\item{method}{character. Method to be used for downloading the data file.}
}
\value{
A data frame containing the refFlat data (see also ?refflat_data)
}
\description{
Fetch refSeq refFlat data from the UCSC genome browser
}
