% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refflat.R
\name{flatten_refflat}
\alias{flatten_refflat}
\title{Flatten refFlat data}
\usage{
flatten_refflat(refflat)
}
\arguments{
\item{refflat}{data frame. The refflat dataset.}

\item{chrom}{character. The chromosome of the slice.}

\item{start}{integer. The start position of the slice.}

\item{end}{integer. The end position of the slice.}
}
\value{
A data frame representing the requested slice of the refflat data.
}
\description{
Flatten refFlat data
}
