% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refflat.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot a gene}
\usage{
plot_gene(name, refflat = refflat_data)
}
\arguments{
\item{name}{character. The name of the gene to plot.}

\item{refflat}{data frame. The refflat dataset.}
}
\description{
Plot a gene
}
