% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refflat.R
\name{slice_refflat}
\alias{slice_refflat}
\title{Slice the refFlat dataset}
\usage{
slice_refflat(chrom, start, end, refflat = refflat_data)
}
\arguments{
\item{chrom}{character. The chromosome of the slice.}

\item{start}{integer. The start position of the slice.}

\item{end}{integer. The end position of the slice.}

\item{refflat}{data frame. The refflat dataset.}
}
\value{
A data frame representing the requested slice of the refflat data.
}
\description{
Slice the refFlat dataset
}
