% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2d_fto.R
\docType{data}
\name{t2d_fto}
\alias{t2d_fto}
\title{Example GWAS data}
\format{A data frame with 3246 rows (SNPs) and 4 variables:
\describe{
  \item{MARKER}{string describing a SNP}
  \item{CHR}{chromosome of the SNP coordinate}
  \item{POS}{position of the SNP coordinate}
  \item{PVALUE}{P-value of the T2D association}
}}
\source{
\url{https://www.nature.com/articles/s41588-018-0332-4}
}
\usage{
t2d_fto
}
\description{
Featuring a T2D signal at the FTO locus
}
\keyword{datasets}
