% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refflat.R
\name{plot_refflat}
\alias{plot_refflat}
\title{Plot refflat}
\usage{
plot_refflat(chrom, start, end, rf = refflat_data, flatten = TRUE,
  arrowhead_length = 0.125, angle = 15, lwd = 2, buffer = 0)
}
\arguments{
\item{chrom}{integer. The chromosome of the plotting interval. (e.g. "chr16")}

\item{start}{integer. The start of the plotting interval.}

\item{end}{integer. The end of the plotting interval.}

\item{flatten}{logical. Flatten genes with multiple transcripts.}

\item{arrowhead_length}{numeric. Length of the edges of the arrow head (in
inches).}

\item{angle}{numeric. Angle from the shaft of the arrow to the edge of the
arrow head.}

\item{lwd}{numeric. Weight of lines.}

\item{refflat}{data frame. The refflat dataset.}
}
\description{
Plot refflat data in an interval
}
